% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_data.R
\name{current_data}
\alias{current_data}
\title{Retrieve current data}
\usage{
current_data(station_id, api_key = NULL, x_api_secret = NULL, as_list = FALSE)
}
\arguments{
\item{station_id}{integer. Station identification code (id).}

\item{api_key}{character. Start date. Use the format \code{day/month/year}. Example: \code{20/09/2024}}

\item{x_api_secret}{character. End date. Use the format \code{day/month/year}. Example: \code{20/09/2024}}

\item{as_list}{logical. If \code{TRUE} the function will return a list of lists, otherwise, a list with tibbles.}
}
\value{
a list of lists or a list of tibbles, depending on the \code{as_list} argument.
}
\description{
Retrieve current data
}
